fix: button emoji picker
